+++ src/decorators/GraphQLResolver.ts
@@ -3,12 +3,18 @@ import { ClassType, ClassTypeResolver } from "../types/decorators";
 
 export function GraphQLResolver(typeFunc: ClassTypeResolver): ClassDecorator;
 export function GraphQLResolver(objectType: ClassType): ClassDecorator;
-export function GraphQLResolver(objectTypeOrTypeFunc: Function): ClassDecorator {
-  const getParentType = objectTypeOrTypeFunc.prototype
-    ? () => objectTypeOrTypeFunc as ClassType
-    : (objectTypeOrTypeFunc as ClassTypeResolver);
-
+export function GraphQLResolver(): ClassDecorator;
+export function GraphQLResolver(objectTypeOrTypeFunc?: Function): ClassDecorator {
   return target => {
+    const getParentType = objectTypeOrTypeFunc
+      ? objectTypeOrTypeFunc.prototype
+        ? () => objectTypeOrTypeFunc as ClassType
+        : (objectTypeOrTypeFunc as ClassTypeResolver)
+      : () => {
+          throw new Error(
+            `No provided object type in '@GraphQLResolver' decorator for class '${target.name}!'`,
+          );
+        };
     MetadataStorage.registerResolver({
       target,
       getParentType,